tsdb: erase pointers to removed mmapped chunks
Go's garbage-collector considers all of a slice's memory to be reachable,
so we should blank out all the pointers beyond the ones we want to keep.

Suppose we had chunks 1,2,3,4,5 and we removed 1-4, then the slice memory
will hold 5,2,3,4,5 and the garbage-collector will not remove chunks 2-3
until they are overwritten by new chunks.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>