remote-write: remove low-value detection of dropped series
This map serves only to decide whether to print a message like
"Dropped sample for series that was not explicitly dropped via relabelling".

This message is opaque and not actionable. The condition can arise from
either a logic error or corruption in the WAL. Corruption should be
visible through other errors during checkpointing.

My suggestion is to save the megabytes that the map needs, and drop the
error message.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>